+++ mosquito-entity/src/main/java/com.softserve.mosquito/entities/Specialization.java
@@ -11,7 +11,7 @@ public final class Specialization implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Byte id;
+    private Long id;
     private String title;
 
     @ManyToMany(mappedBy = "specializations")
@@ -23,16 +23,16 @@ public final class Specialization implements Serializable {
         this.title = title;
     }
 
-    public Specialization(Byte id, String title) {
+    public Specialization(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public Byte getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Byte id) {
+    public void setId(Long id) {
         this.id = id;
     }
